Move payment relationships into junction table
In its previous form, the database only supported many-to-one
relationships between payments (`dfl_orders_payments`) and donations
(`donations`) or subscriptions (`dfl_users_subscriptions`). With mass
sub gifting, it's now possible that a single payment is linked to
multiple subscriptions, so we need to create a junction table,
`dfl_payments_purchases`, to hold the associations.