Move sub creation to `subscriptionProcess()`
Previously, we created a sub with `NEW` status in
`subscriptionCreate()`, then updated its status to `ACTIVE` or `PENDING`
in `subscriptionProcess()`. When `subscriptionProcess()` handles the
entirety of this process, we (1) no longer have to worry about malicious
users attacking `/subscription/create` to flood the database with new
sub records, and (2) can implement mass gifting without needing to pass
up to 100 sub ID's between routes.