Fix sub details in query string vulnerability
Creating a subscription in `/process` rather than `/create` introduced
an exploit where a specially-crafted request allows a malicious user to
purchase a mass gift for the price of a single sub.

We address the problem by no longer passing subscription parameters to
the endpoint in a query string. Instead, all the details of the
subscription are extracted directly from the EC checkout details, which
is validated and can't be forged. This includes the giftee, which is
JSON encoded and stored in the `Custom` field.

As a bonus, we can now display the giftee on the `/complete` page.