Create membership-related database tables
We use a composite primary key for `youtube_memberships` to elegantly
handle scenarios where the user upgrades/downgrades their membership. If
a record with the same `memberChannelId` and `creatorChannelId` exist,
we simply upgrade `membershipLevelId`.

`channelId` is the primary key for `users_youtube_channel_ids` to handle
a situations where two users try to log into the same account. The first
user would lose their ability to sync their membership.

We need to replace all occurrences of `youtube` with
`youtubebroadcaster` in `dfl_users_auth` because the handler's name was
updated in an earlier commit.