Add display name config option
All we need to embed a Twitch livestream is a channel name, a
user-readable value we can display in the host pill as-is. With YouTube
livestreams, we need a channel ID, which doesn't look nearly as nice.

Supplying a "display name" to be used in the host pill across all
streams seems to be the simplest solution to this problem.