Fix issue with embedding other Twitch streams
Embedding an "outside" Twitch stream while watching a YouTube embed
yields an error about the embed being misconfigured. This is because
`parent` isn't set in the embed URL's query string.

We solve the issue by moving `twitchParents` up from a `stream` option
to a general `embed` option in our config file and applying it to all
Twitch streams.