Enforce karma and account age requirements for reddit logins (#257)
* Add config options for min reddit karma and age

* Create reddit service func to get basic user info

* Validate accounts used for OAuth connections

Only new connections are validated. For the time being, the only
requirements are minimum karma and account age for reddit accounts.

* Fix `/login` redirect handling when logged in

If an exception is thrown when a user tries to connect a new auth,
they're redirected to `/login`, then immediately to `/profile`, and
shown an error message about already being signed in. Showing the
message from the original exception and redirecting to
`/profile/authentication` is a better approach.

* Version bump