Improve YouTube livestream integration (#263)
* Create new task for YouTube-related operations

* Temporarily disable original YouTube vid feed task

* Return recent YouTube uploads as-is

No longer filter out non-public videos or modify the publish date.

* Fetch videos using ID's from playlist items

The `playlistItem` resource doesn't contain all the data we need to
process everything (whether or not the video is a live broadcast,
specifically).

* Update recent YT videos within new YouTube task

* Delete old YouTube uploads task

* Fix bug with using wrong ID when getting videos

The ID of a playlist item is different than the underlying video's ID.

* Fix bug with rendering videos view

We don't store the entire response from YouTube, only the videos
themselves. Using the `items` key is no longer necessary.

* Cache completed YT broadcast VODs

* Normalize uploads and broadcasts before caching

* Cache info on latest YT livestream

* Fix bug with fetching/accessing video ID

* Correctly access YT stream start and end times

* Only access viewers if the stream is live

If the stream isn't live, the value doesn't exit.

* Remove `preview`

Preview thumbnails aren't used on the Bigscreen page.

* Enable autoplay and inline playback for YT embeds

* Tweak approach to managing multiple embeds

The new approach offers less flexibility (you can no longer insert any
number of streams to toggle between), but sets the foundation for
dynamically updating the YouTube livestream URL on the client.

* Track stream info for each embed separately

This allows us to better convey to users when one stream is online and
the other isn't.

* Return to YouTube auth page after auth is complete

* Fix issue with showing hosted channel on homepage

* Correctly calculate duration for finished streams

* Rename some vars/funcs for clarity/consistency

* Fix bugs with stream info API response

We should calculate the MD5 hash of the entire response to ensure the
ETag header works as intended.

* Update YT stream embed automatically

This ensures the YouTube stream starts playing back automatically when
the streamer goes live if the user has the YouTube embed up. However,
this functionality relies on a response from `/api/info/stream`, which
is only pinged every 90 seconds.

* Embed current YT stream when Bigscreen loads

* Version bump